requirements.txt
einops==0.4.1
imageio==2.22.1
mediapipe==0.8.11
numpy==1.23.2
Pillow==9.2.0
scikit_image==0.19.3
scipy==1.9.1
skimage==0.0
torch==1.12.1
torchvision==0.13.1
